+++ lib/easypost/pickup.rb
@@ -25,11 +25,6 @@ class EasyPost::Pickup < EasyPost::Resource
     self
   end
 
-  # Retrieve a list of all Pickup objects.
-  def self.all(_filters = {}, _api_key = nil)
-    raise NotImplementedError.new('Pickup.all not implemented.')
-  end
-
   # Get the lowest rate of a Pickup (can exclude by having `'!'` as the first element of your optional filter lists).
   def lowest_rate(carriers = [], services = [])
     EasyPost::Util.get_lowest_object_rate(self, carriers, services, 'pickup_rates')