+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Added function to retrieve all pickups, accessible via `EasyPost::Pickup.all`
 - Added payload functions `retrieve_all_payloads` and `retrieve_payload` to retrieve all payloads or a specific payload for an event.
 
 ## v4.10.0 (2023-01-11)